build restructure; configs restructure; added telegram apis into tg package; added main script
