+++ README.md
@@ -9,10 +9,17 @@ Steps on build.
 ### Install dependencies
 
 From the root dir:
+Firstly, it's required to install dependencies and prepare scripts.
 
 ```sh
 pip install .
 # or if it does not work, try `python -m pip install .`
+```
+
+### Run project
+
+```sh
+main
 ```
 
 ### Run tests
@@ -29,4 +36,3 @@ pytest
 - Person 3
 - Person 4
 - Person 5
--
