+++ app/entry/__main__.py
@@ -1,7 +1,5 @@
 import sys
-
-def main(args = None):
-    pass
+from app.entry.entry import *
 
 if __name__ == "__main__":
     sys.exit(main())
