--- config.py
@@ -1,1 +1,0 @@
-API_TOKEN = ''
