+++ pyproject.toml
@@ -6,8 +6,10 @@ description = "todo"
 requires-python = ">=3.8"
 dependencies = [
     "psycopg2 >= 2.9.10",
-    "pytest >= 8.3.0"
+    "pytest >= 8.3.0",
+    "aiogram >= 3.14.0"
 ]
+readme = "README.md"
 
 [tool.pytest.ini_options]
 minversion = "6.0"
@@ -15,3 +17,14 @@ addopts = "-rfE"
 testpaths = [
     "test"
 ]
+
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[tool.hatch.build.targets.wheel]
+packages = ["app/api", "app/configs", "app/db", "app/entry", "app/models", "app/tg"]
+
+[project.scripts]
+main = "app.entry.entry:main"
+
